Struct google_cloudresourcemanager1::Binding

pub struct Binding {
    pub role: Option<String>,
    pub members: Option<Vec<String>>,
}

Associates members with a role.

This type is not used in any activity, and only used as part of another schema.

Fields

role

Role that is assigned to members. For example, roles/viewer, roles/editor, or roles/owner. Required

members

Specifies the identities requesting access for a Cloud Platform resource. members can have the following values: * allUsers: A special identifier that represents anyone who is on the internet; with or without a Google account. * allAuthenticatedUsers: A special identifier that represents anyone who is authenticated with a Google account or a service account. * user:{emailid}: An email address that represents a specific Google account. For example, alice@gmail.com or joe@example.com. * serviceAccount:{emailid}: An email address that represents a service account. For example, my-other-app@appspot.gserviceaccount.com. * group:{emailid}: An email address that represents a Google group. For example, admins@example.com. * domain:{domain}: A Google Apps domain name that represents all the users of that domain. For example, google.com or example.com.