Struct google_cloudresourcemanager1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudResourceManager hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_cloudresourcemanager1 as cloudresourcemanager1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use cloudresourcemanager1::CloudResourceManager;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CloudResourceManager::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `get_iam_policy(...)`, `list(...)`, `set_iam_policy(...)`, `test_iam_permissions(...)`, `undelete(...)` and `update(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn get(&self, project_id: &str) -> ProjectGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the Project identified by the specified project_id (for example, my-project-123). The caller must have read permissions for this Project.

Arguments

projectId - The Project ID (for example, my-project-123). Required.

fn undelete(&self, request: UndeleteProjectRequest, project_id: &str) -> ProjectUndeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Restores the Project identified by the specified project_id (for example, my-project-123). You can only use this method for a Project that has a lifecycle state of DELETE_REQUESTED. After deletion starts, as indicated by a lifecycle state of DELETE_IN_PROGRESS, the Project cannot be restored. The caller must have modify permissions for this Project.

Arguments

request - No description provided.

projectId - The project ID (for example, foo-bar-123). Required.

fn set_iam_policy(&self, request: SetIamPolicyRequest, resource: &str) -> ProjectSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the IAM access control policy for the specified Project. Replaces any existing policy. The following constraints apply when using setIamPolicy(): + Project currently supports only user:{emailid} and serviceAccount:{emailid} members in a Binding of a Policy. + To be added as an owner, a user must be invited via Cloud Platform console and must accept the invitation. + Members cannot be added to more than one role in the same policy. + There must be at least one owner who has accepted the Terms of Service (ToS) agreement in the policy. Calling setIamPolicy() to to remove the last ToS-accepted owner from the policy will fail. + Calling this method requires enabling the App Engine Admin API. Note: Removing service accounts from policies or changing their roles can render services completely inoperable. It is important to understand how the service account is being used before removing or updating its roles.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. resource is usually specified as a path, such as projects/*project*/zones/*zone*/disks/*disk*. The format for the path specified in this value is resource specific and is specified in the setIamPolicy documentation.

fn test_iam_permissions(&self, request: TestIamPermissionsRequest, resource: &str) -> ProjectTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified Project.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. resource is usually specified as a path, such as projects/*project*/zones/*zone*/disks/*disk*. The format for the path specified in this value is resource specific and is specified in the testIamPermissions documentation.

fn update(&self, request: Project, project_id: &str) -> ProjectUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the attributes of the Project identified by the specified project_id (for example, my-project-123). The caller must have modify permissions for this Project.

Arguments

request - No description provided.

projectId - The project ID (for example, my-project-123). Required.

fn get_iam_policy(&self, request: GetIamPolicyRequest, resource: &str) -> ProjectGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the IAM access control policy for the specified Project. Permission is denied if the policy or the resource does not exist.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. resource is usually specified as a path, such as projects/*project*/zones/*zone*/disks/*disk*. The format for the path specified in this value is resource specific and is specified in the getIamPolicy documentation.

fn delete(&self, project_id: &str) -> ProjectDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Marks the Project identified by the specified project_id (for example, my-project-123) for deletion. This method will only affect the Project if the following criteria are met: + The Project does not have a billing account associated with it. + The Project has a lifecycle state of ACTIVE. This method changes the Project's lifecycle state from ACTIVE to DELETE_REQUESTED. The deletion starts at an unspecified time, at which point the lifecycle state changes to DELETE_IN_PROGRESS. Until the deletion completes, you can check the lifecycle state checked by retrieving the Project with GetProject, and the Project remains visible to ListProjects. However, you cannot update the project. After the deletion completes, the Project is not retrievable by the GetProject and ListProjects methods. The caller must have modify permissions for this Project.

Arguments

projectId - The Project ID (for example, foo-bar-123). Required.

fn list(&self) -> ProjectListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists Projects that are visible to the user and satisfy the specified filter. This method returns Projects in an unspecified order. New Projects do not necessarily appear at the end of the list.